Paginates in order to keep performance of the app fast. Also allows for all resources picked up to be displayed to the user

+++ Groundskeeper.py
@@ -20,15 +20,31 @@ def check_expiration_tags(resource_client):
     expiration_tag_key = 'ExpirationDate'
     threshold_date = datetime.utcnow().date()
 
-    resources = resource_client.resources.list()
-    for resource in resources:
-        if resource.tags and expiration_tag_key in resource.tags:
-            expiration_date = datetime.strptime(resource.tags[expiration_tag_key], '%Y-%m-%d').date()
-            if expiration_date <= threshold_date:
-                resource_name = resource.id.split('/')[-1]
-                delete_resource(resource_client, resource.id, resource_name)  # Pass resource.id and resource_name here
+    # List to store resources for deletion
+    resources_for_deletion = []
 
+    # Fetch resources by page
+    resources_pages = resource_client.resources.list().by_page()
 
+    # Iterate over pages of resources
+    for resources_page in resources_pages:
+        # Iterate over resources on the current page
+        for resource in resources_page:
+            if resource.tags and expiration_tag_key in resource.tags:
+                expiration_date = datetime.strptime(resource.tags[expiration_tag_key], '%Y-%m-%d').date()
+                if expiration_date <= threshold_date:
+                    resource_name = resource.id.split('/')[-1]
+                    # Check if it's dry-run mode and add resource to the list
+                    if dry_run:
+                        resources_for_deletion.append(resource_name)
+                    else:
+                        delete_resource(resource_client, resource.id, resource_name)
+
+    # Print resources picked up for deletion in dry-run mode
+    if dry_run:
+        print("Resources that would be deleted:")
+        for resource_name in resources_for_deletion:
+            print(f"- {resource_name}")
 
 def delete_resource(resource_client, resource_id, resource_name):
     if dry_run:
@@ -40,9 +56,6 @@ def delete_resource(resource_client, resource_id, resource_name):
             print(f"Resource {resource_name} has been deleted.")
         except Exception as e:
             print(f"Failed to delete resource {resource_name}: {e}")
-
-
-
 
 if __name__ == "__main__":
     # |||IMPORTANT|||
